fixed api routes with headers user validation

+++ src/controllers/api/categoriesController.js
@@ -7,7 +7,9 @@ import Category from "../../models/Category.js";
  */
 export const getCategories = async (req, res) => {
   try {
-    const categories = await Category.query();
+    const categories = await Category.query().where({
+      user_id: req.user.id,
+    });
     res.status(200).json(categories);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -26,6 +28,13 @@ export const getCategory = async (req, res) => {
     if (!category) {
       return res.status(404).json({ message: "Category not found!" });
     }
+
+    if (category.user_id !== req.user.id) {
+      return res
+        .status(403)
+        .json({ message: "You do not have permission to view this category." });
+    }
+
     res.status(200).json(category);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -47,6 +56,7 @@ export const createCategory = async (req, res) => {
     if (!category) {
       category = await Category.query().insert({
         link,
+        user_id: req.user.id,
         ...req.body,
       });
     }
@@ -70,6 +80,12 @@ export const updateCategory = async (req, res) => {
       return res.status(404).json({ message: "Category not found!" });
     }
 
+    if (category.user_id !== req.user.id) {
+      return res.status(403).json({
+        message: "You do not have permission to update this category.",
+      });
+    }
+
     await category.$query().patch(req.body);
 
     res.status(200).json(category);
@@ -90,6 +106,14 @@ export const deleteCategory = async (req, res) => {
 
     if (!category) {
       return res.status(404).json({ message: "Category not found!" });
+    }
+
+    if (category.user_id !== req.user.id) {
+      return res
+        .status(403)
+        .json({
+          message: "You do not have permission to delete this category.",
+        });
     }
 
     await category.$query().delete();
